use v8

+++ app/http/endpoints/api/guilds.go
@@ -42,7 +42,7 @@ func GetGuilds(ctx *gin.Context) {
 			fakeGuild := guild.Guild{
 				Id:          g.GuildId,
 				Owner:       g.Owner,
-				Permissions: int(g.UserPermissions),
+				Permissions: uint64(g.UserPermissions),
 			}
 
 			if g.Owner {
